+++ modules/photons_canvas/animations/infrastructure/animation.py
@@ -40,7 +40,7 @@ class Animation:
 
     @hp.memoized_property
     def ticker(self):
-        return hp.ATicker(self.every, max_time=self.num_seconds)
+        return hp.ATicker(self.every, max_time=self.num_seconds, min_wait=0.01)
 
     async def stream(self, animation_state):
         stop_fut = hp.ChildOfFuture(self.final_future)